lastupdate timestamp data loaded by rrd.bufferValue types the data as an Integer, rrdtool can handle sub-second timestamp data (specifically useful handling drift when updating the rrd every second).  This should instead be a float.
This fixes the following:
  http://code.google.com/p/pyrrd/issues/detail?id=32

+++ pyrrd/rrd.py
@@ -163,7 +163,7 @@ class RRD(mapper.RRDMapper):
         """
         values = ':'.join([unicode(x) for x in values])
         self.values.append((timeOrData, values))
-        self.lastupdate = int(unicode(timeOrData).split(":")[0])
+        self.lastupdate = float(unicode(timeOrData).split(":")[0])
 
     # for backwards compatibility
     bufferValues = bufferValue
